Move filter dropdown to seperate component
