+++ src/components/Filter.tsx
@@ -1,61 +1,22 @@
 import '../styles/Filter.css'
-import { useEffect, useState } from "react";
-import CommitteeButton from "./CommitteeButton";
-import { committeeLogos } from "./LinkModule";
-import {
-    sites,
-    defaultCommitteeHash,
-    CommitteeHashType,
-    CommitteeDict,
-    PageInfo,
-} from "../siteContent";
+import { PageInfo } from "../siteContent";
+import FilterDropdown from './FilterDropdown';
+import { useState } from 'react';
 
 interface FilterProps {
     onChange: (links: [string, PageInfo[]][]) => void;
 }
 
 export default function Filter(props: FilterProps) {
-    //make a list of committee buttons
-    const [displayedCommittees, setDisplayedCommittees] =
-        useState(defaultCommitteeHash);
-    const committees = Object.keys(sites).map((site) => (
-        <CommitteeButton
-            key={site}
-            onClick={() => {
-                setDisplayedCommittees({
-                    ...displayedCommittees,
-                    [site]: !displayedCommittees[site as keyof CommitteeHashType],
-                })
-            }
-            }
-            on={displayedCommittees[site as keyof CommitteeHashType]}
-            imgSrc={committeeLogos[site as keyof CommitteeDict]}
-            site={site}
-        />
-    ));
-    useEffect(() => {
-        //SiteCheck holds all filtered sites
-        const siteCheck = Object.entries(sites)
-            .filter((site) => {
-                const initiative = site[0];
-                //only return the sites if they are currently being filtered
-                return displayedCommittees[initiative as keyof CommitteeHashType];
-            });
-        props.onChange(siteCheck);
-    }, [displayedCommittees, props]);
-
+    const [dropdownOpen,setDropdownOpen] = useState(false);
     return (
         <div>
-            <p>
-                Filter By:
-                <div className="button-list">{committees}</div>
-            </p>
-            <button
-                onClick={() => {
-                    setDisplayedCommittees(defaultCommitteeHash);
-                }}>
-                Reset All
-            </button>
+            <div>
+                <input placeholder='Search sites...' />
+                <button>Search</button>
+                <button onClick={() => setDropdownOpen(dropdownOpen => !dropdownOpen)}>Filter</button>
+            </div>
+            {dropdownOpen && <FilterDropdown onChange={(sites) => props.onChange(sites)} />}
         </div>
     )
 }
